Fixed columns types for amounts

+++ database/migrations/2024_02_24_165000_create_payu_transactions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->uuid('id')->primary();
             $table->string('payu_order_id', 50);
             $table->text('link');
-            $table->unsignedInteger('amount')->default(0);
+            $table->unsignedBigInteger('amount')->default(0);
             $table->json('payload');
             $table->unsignedTinyInteger('status')->default(PaymentStatus::INITIALIZED);
             $table->string('method_id')->nullable();
